Added basics for roles

+++ src/main/java/org/project/controller/BookController.java
@@ -5,6 +5,7 @@ import jakarta.servlet.http.HttpSession;
 import org.project.model.Book;
 import org.project.model.Rating;
 import org.project.model.Series;
+import org.project.model.User;
 import org.project.repository.BookRepository;
 import org.project.repository.SeriesRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -60,6 +61,11 @@ public class BookController {
             @RequestParam(required = false, defaultValue = "") String function,
             @RequestParam(required = false) String variable,
             Model model, HttpSession session) {
+
+        User user = (User) session.getAttribute("currentUser");
+        if (user == null || !user.getIsOwner()){
+            return "redirect:/";
+        }
 
         Iterable<Book> bookList = null;
 
@@ -148,7 +154,12 @@ public class BookController {
     }
 
     @PostMapping("/add-book")
-    public String createBook(@Valid @ModelAttribute Book book, BindingResult bindingResult, Model model, @RequestParam ("pictureUpload") MultipartFile file, @RequestParam("seriesName")String seriesName){
+    public String createBook(@Valid @ModelAttribute Book book, BindingResult bindingResult, Model model, @RequestParam ("pictureUpload") MultipartFile file, @RequestParam("seriesName")String seriesName, HttpSession session){
+        User user = (User) session.getAttribute("currentUser");
+        if (user == null || !user.getIsOwner()){
+            throw new RuntimeException("Invalid permissions");
+        }
+
         if(!file.isEmpty()){
             try{
                 byte[] bytes = file.getBytes();
@@ -185,13 +196,22 @@ public class BookController {
     }
 
     @PostMapping("/delete-book/{ISBN}")
-    public String deleteBook(@PathVariable long ISBN){
+    public String deleteBook(@PathVariable long ISBN, HttpSession session){
+        User user = (User) session.getAttribute("currentUser");
+        if (user == null || !user.getIsOwner()){
+            throw new RuntimeException("Invalid permissions");
+        }
         bookRepo.deleteById(ISBN);
         return "redirect:/get-book-list";
     }
 
     @GetMapping("/edit-book/{ISBN}")
-    public String editBook(@PathVariable long ISBN, Model model){
+    public String editBook(@PathVariable long ISBN, Model model, HttpSession session){
+        User user = (User) session.getAttribute("currentUser");
+        if (user == null || !user.getIsOwner()){
+            throw new RuntimeException("Invalid permissions");
+        }
+
         Book book = bookRepo.findByISBN(ISBN);
         model.addAttribute("series", seriesRepo.findAll());
         model.addAttribute("book", book);
@@ -200,7 +220,12 @@ public class BookController {
     }
 
     @PostMapping("/update-book")
-    public String updateBook(@ModelAttribute Book book, @RequestParam ("pictureUpload") MultipartFile file, @RequestParam("seriesName")String seriesName){
+    public String updateBook(@ModelAttribute Book book, @RequestParam ("pictureUpload") MultipartFile file, @RequestParam("seriesName")String seriesName, HttpSession session){
+        User user = (User) session.getAttribute("currentUser");
+        if (user == null || !user.getIsOwner()){
+            throw new RuntimeException("Invalid permissions");
+        }
+
         if(!file.isEmpty()){
             try{
                 byte[] bytes = file.getBytes();
